0.4.0 rename to pyexiftool and make sure tests pass
